state - renames 'redux' directory to 'state'
